+++ CHANGELOG.md
@@ -7,11 +7,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 This project does not follow SemVer, since modules are independent of each other; thus, SemVer does not make sense. Changes are grouped per module.
 
 ## Unreleased
+### Cloudtrail
+* New module which allows the creation of management and data cloudtrail
+
 ### hashicorp/aws
 - Upgrade to v5. [#271](https://github.com/dbl-works/terraform/pull/271)
 
 ### cloudwatch-kinesis
 - hashicorp/aws: Drop support for v4 and add support for v5.[#271](https://github.com/dbl-works/terraform/pull/271)
+
 
 ## [v2023.11.08]
 ### script/database-roles